Start refactoring code inside net
