+++ .gitignore
@@ -82,11 +82,8 @@ test/dns/error
 test/dns/query
 test/fixtures/GeoIP.dat
 test/fixtures/GeoIPASNum.dat
-test/http/client
-test/http/funcs
 test/http/parse_url
 test/http/request
-test/http/request_serializer
 test/http/response_parser
 test/mlabns/mlabns
 test/ndt/messages
