+++ include/measurement_kit/net.hpp
@@ -2,6 +2,7 @@
 #ifndef MEASUREMENT_KIT_NET_HPP
 #define MEASUREMENT_KIT_NET_HPP
 #include <measurement_kit/net/buffer.hpp>
+#include <measurement_kit/net/connect.hpp>
 #include <measurement_kit/net/error.hpp>
 #include <measurement_kit/net/transport.hpp>
 #endif
