+++ include/measurement_kit/net/transport.hpp
@@ -4,22 +4,14 @@
 #ifndef MEASUREMENT_KIT_NET_TRANSPORT_HPP
 #define MEASUREMENT_KIT_NET_TRANSPORT_HPP
 
-#include <functional>
-#include <measurement_kit/common.hpp>
 #include <measurement_kit/dns.hpp>
 #include <measurement_kit/net/buffer.hpp>
 #include <stddef.h>
-#include <string>
-#include <vector>
 
 // Forward declaration
 struct bufferevent;
 
 namespace mk {
-
-// Forward declaration
-class Error;
-
 namespace net {
 
 class Transport {
@@ -56,37 +48,6 @@ class Transport {
     virtual std::string socks5_address() = 0;
     virtual std::string socks5_port() = 0;
 };
-
-struct ResolveHostnameResult {
-    bool inet_pton_ipv4 = false;
-    bool inet_pton_ipv6 = false;
-
-    Error ipv4_err;
-    dns::Message ipv4_reply;
-    Error ipv6_err;
-    dns::Message ipv6_reply;
-
-    std::vector<std::string> addresses;
-};
-
-struct ConnectResult : public ErrorContext {
-    ResolveHostnameResult resolve_result;
-    std::vector<Error> connect_result;
-    bufferevent *connected_bev = nullptr;
-};
-
-void connect(std::string address, int port,
-             Callback<Error, Var<Transport>> callback,
-             Settings settings = {},
-             Var<Logger> logger = Logger::global(),
-             Var<Reactor> reactor = Reactor::global());
-
-using ConnectManyCb = Callback<Error, std::vector<Var<Transport>>>;
-
-void connect_many(std::string address, int port, int num,
-        ConnectManyCb callback, Settings settings = {},
-        Var<Logger> logger = Logger::global(),
-        Var<Reactor> reactor = Reactor::global());
 
 void write(Var<Transport> txp, Buffer buf, Callback<Error> cb);
